+++ src/main/java/pl/wat/db/repository/event/EventPictureRepository.java
@@ -3,5 +3,5 @@ package pl.wat.db.repository.event;
 import org.springframework.data.repository.CrudRepository;
 import pl.wat.db.domain.event.EventPicture;
 
-public interface EventPictureRepository extends CrudRepository<EventPicture, Integer> {
+public interface EventPictureRepository extends CrudRepository<EventPicture, Long> {
 }
