+++ src/main/java/pl/wat/db/repository/event/LocalizatonRepository.java
@@ -3,5 +3,5 @@ package pl.wat.db.repository.event;
 import org.springframework.data.repository.CrudRepository;
 import pl.wat.db.domain.event.Localization;
 
-public interface LocalizatonRepository extends CrudRepository<Localization, Integer> {
+public interface LocalizatonRepository extends CrudRepository<Localization, Long> {
 }
