+++ src/main/java/pl/wat/db/repository/personality/CategoryAttributeRepository.java
@@ -3,5 +3,5 @@ package pl.wat.db.repository.personality;
 import org.springframework.data.repository.CrudRepository;
 import pl.wat.db.domain.personality.CategoryAttribute;
 
-public interface CategoryAttributeRepository extends CrudRepository<CategoryAttribute, Integer> {
+public interface CategoryAttributeRepository extends CrudRepository<CategoryAttribute, Long> {
 }
