+++ src/main/java/pl/wat/db/repository/personality/PersonalityCategoryRepository.java
@@ -3,5 +3,5 @@ package pl.wat.db.repository.personality;
 import org.springframework.data.repository.CrudRepository;
 import pl.wat.db.domain.personality.PersonalityCategory;
 
-public interface PersonalityCategoryRepository extends CrudRepository<PersonalityCategory, Integer> {
+public interface PersonalityCategoryRepository extends CrudRepository<PersonalityCategory, Long> {
 }
