+++ src/main/java/pl/wat/logic/service/utils/ChatRequest.java
@@ -3,8 +3,8 @@ package pl.wat.logic.service.utils;
 import pl.wat.logic.dto.conversation.ConversationDTO;
 
 public class ChatRequest {
-    public int conversationId;
+    public Long conversationId;
     public ConversationDTO conversation;
-    public int messageId;
+    public Long messageId;
     public String messageText;
 }
